requirements.txt
# Core dependencies
pyyaml==6.0.1
pathlib2==2.3.7
typing-extensions==4.8.0

# Security
cryptography==41.0.7
validators==0.22.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code quality
flake8==6.1.0
black==23.11.0
mypy==1.7.1

# Development
pre-commit==3.5.0

# Utilities
click==8.1.7
colorama==0.4.6
tqdm==4.66.1

# Machine Learning (for TF-IDF implementation)
scikit-learn==1.3.2